requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth
PyMuPDF
python-docx
faiss-cpu
groq
pydantic
numpy
requests
